+++ src/utils.ts
@@ -1,3 +1,7 @@
+export function notNull<T>(obj: null | T): obj is T {
+  return obj !== null;
+}
+
 export function ordinalOf(n: number): string {
   const j = n % 10;
   const k = n % 100;
@@ -6,8 +10,4 @@ export function ordinalOf(n: number): string {
   if (j === 2 && k !== 12) return `${s}nd`;
   if (j === 3 && k !== 13) return `${s}rd`;
   return `${s}th`;
-}
-
-export function notNull<T>(obj: null | T): obj is T {
-  return obj !== null;
 }